+++ app/models/user.rb
@@ -7,7 +7,7 @@ class User < ApplicationRecord
   devise :database_authenticatable, :registerable,
          :recoverable, :rememberable, :validatable
 
-  has_many :identities
+  has_many :identities, dependent: :delete_all
   has_many :payments
 
   after_create :send_welcome_email
