--- app/views/admin/admin/index.html.erb
@@ -1,11 +1,0 @@
-<h2>Admin</h2>
-
-<h3>
-<%= @users.count %>
-<%= "user".pluralize(@users.count) %>
-<% if @users.count > 1 %>
-  🥳
-<% else %>
-  😢
-<% end %>
-</h3>
