+++ app/views/admin/users/show.html.erb
@@ -1,0 +1,2 @@
+<h2><%= @user.email %></h2>
+created at <time><%= @user.created_at %></time>
