+++ db/migrate/20230415145356_add_is_admin_to_users.rb
@@ -1,5 +1,5 @@
 class AddIsAdminToUsers < ActiveRecord::Migration[7.0]
   def change
-    add_column :users, :is_admin, :boolean
+    add_column :users, :is_admin, :boolean, default: false, null: false
   end
 end
