+++ db/schema.rb
@@ -46,7 +46,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_04_15_145356) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.string "uid"
-    t.boolean "is_admin"
+    t.boolean "is_admin", default: false, null: false
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
